% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_column.R
\name{dual_column}
\alias{dual_column}
\title{Render two input Rmd files as equal width columns.}
\usage{
dual_column(left.is.text = TRUE, left, right, img,
  working_dir = getwd())
}
\arguments{
\item{left.is.text}{True if the content of the left file is text only.}

\item{left}{The path to the local Rmd file for the left column.}

\item{right}{The path to the local Rmd file for the right column.}

\item{img}{path to the local image file to insert into the right column.}

\item{working_dir}{The working directory to base relative paths from}
}
\value{
Dual column markdown correctly rendered using knitr irrespective of the output format.
}
\description{
Note this function is rather limited: it is expected for use inside R chunks with the results='asis' and warning=FALSE options specified.
}
\examples{
dual_column (left='left-file.Rmd', right='right-file.Rmd', img='image-file.jpg', working_dir='./')
dual_column (left.is.text=FALSE, left='left-file.Rmd', right='right-file.Rmd', img='image-file.jpg')
}
